# cmd/settlr/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Settlr API",
    "version": "1.0.0",
    "description": "Free/open-source Splitwise alternative. Integer cents ledger, Argon2id, JWT refresh rotation, cursor pagination."
  },
  "servers": [
    {
      "url": "https://settlrapi.theswissknife.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/me/balances": {
      "get": {
        "summary": "My balances",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/expenses/{id}/comments": {
      "get": {
        "summary": "List comments",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Add comment",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/comments/{id}": {
      "delete": {
        "summary": "Delete comment",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/expenses/{id}/attachments": {
      "get": {
        "summary": "List attachments",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "Upload attachment",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/attachments/{id}": {
      "delete": {
        "summary": "Delete attachment",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/groups/{id}/balances": {
      "get": {
        "summary": "Balances",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/groups/{id}/debts": {
      "get": {
        "summary": "Debts",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}